reflexion-agent/pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Anthony Kugel",email = "anthonykugel@gmail.co"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "isort (>=5.13.2,<6.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "langchain-openai (>=0.3.0,<0.4.0)",
    "langgraph (>=0.2.62,<0.3.0)",
    "langchain-community (>=0.3.14,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
